+++ CHANGELOG.md
@@ -1,4 +1,3 @@
-
 # Changelog
 
 All notable changes to this project will be documented in this file.
@@ -11,10 +10,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release of Tabs Per Project plugin
 
-### Security
-
 ## [1.0.0] - 2025-06-02
 
 ### Added
 - Initial release of Tabs Per Project plugin
 - Basic functionality for managing tabs per project
+
+## [1.1.0] - 2025-06-04
+
+### Changed
+- Close button in the project header now uses a standard icon for a more modern look.
+- Settings UI improved for better usability.
+
+### Added
+- Added a "Reset Style" button in the settings to restore all style-related options to their default values.
+- More customization options for project and module header styles (font, color, text case).
+- When pressing the middle mouse button on a file, the file will now close